docs(Forms): move more-fields to feature-fields and base-fields
Move Password from more-fields/ to feature-fields/.
Move Slider and Upload from more-fields/ to base-fields/.
Remove the more-fields/ subdirectory and its listing page.
Update all breadcrumbs, imports, links, and glob patterns.
