+++ packages/dnb-design-system-portal/src/docs/uilib/about-the-lib/releases/eufemia/v11-info.mdx
@@ -2795,7 +2795,7 @@ See docs about changed [Dropdown properties](/uilib/about-the-lib/releases/eufem
 - Replace `search_content` with `searchContent`.
 - Replace `class_name` with `className`.
 
-### [Field.Password](/uilib/extensions/forms/feature-fields/more-fields/Password/)
+### [Field.Password](/uilib/extensions/forms/feature-fields/Password/)
 
 #### Properties
