+++ packages/dnb-design-system-portal/src/docs/uilib/components/input/demos.mdx
@@ -79,7 +79,7 @@ You can check out the [Field.String](/uilib/extensions/forms/base-fields/String/
 
 The password component has to ensure that there is still room for password managers to inject the input with their UX functionality.
 
-Read more about it in [Field.Password](/uilib/extensions/forms/feature-fields/more-fields/Password/).
+Read more about it in [Field.Password](/uilib/extensions/forms/feature-fields/Password/).
 
 <VisibleWhenVisualTest>
   <InputScreenshotTests />
