+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/Value/Upload/info.mdx
@@ -13,7 +13,7 @@ render(<Value.Upload />)
 
 `Value.Upload` is a value component for displaying a list of files.
 
-There is a corresponding [Field.Upload](/uilib/extensions/forms/feature-fields/more-fields/Upload) component.
+There is a corresponding [Field.Upload](/uilib/extensions/forms/base-fields/Upload) component.
 
 ```jsx
 import { Value } from '@dnb/eufemia/extensions/forms'
