+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/base-fields/Slider.mdx
@@ -1,6 +1,7 @@
 ---
 title: 'Field.Slider'
 description: '`Field.Slider` is a wrapper component for the Slider to make it easier to use inside a form.'
+componentType: 'base-input'
 showTabs: true
 theme: ['sbanken', 'carnegie']
 tabs:
@@ -15,16 +16,14 @@ tabs:
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
-  - text: Feature fields
-    href: /uilib/extensions/forms/feature-fields/
-  - text: More feature fields...
-    href: /uilib/extensions/forms/feature-fields/more-fields/
+  - text: Base fields
+    href: /uilib/extensions/forms/base-fields/
   - text: Slider
-    href: /uilib/extensions/forms/feature-fields/more-fields/Slider/
+    href: /uilib/extensions/forms/base-fields/Slider/
 ---
 
-import Info from 'Docs/uilib/extensions/forms/feature-fields/more-fields/Slider/info'
-import Demos from 'Docs/uilib/extensions/forms/feature-fields/more-fields/Slider/demos'
+import Info from 'Docs/uilib/extensions/forms/base-fields/Slider/info'
+import Demos from 'Docs/uilib/extensions/forms/base-fields/Slider/demos'
 
 <Info />
 <Demos />
