+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/base-fields/Slider/Examples.tsx
@@ -1,5 +1,5 @@
 import { Flex, HelpButton, P } from '@dnb/eufemia/src'
-import ComponentBox from '../../../../../../../shared/tags/ComponentBox'
+import ComponentBox from '../../../../../../shared/tags/ComponentBox'
 import { Field, Form, Value } from '@dnb/eufemia/src/extensions/forms'
 
 export const BasicUsage = () => {
