+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/base-fields/Slider/info.mdx
@@ -16,7 +16,7 @@ render(<Field.Slider />)
 ## Relevant links
 
 - [Source code](https://github.com/dnbexperience/eufemia/tree/main/packages/dnb-eufemia/src/extensions/forms/Field/Slider)
-- [Docs code](https://github.com/dnbexperience/eufemia/tree/main/packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/feature-fields/more-fields/Slider)
+- [Docs code](https://github.com/dnbexperience/eufemia/tree/main/packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/base-fields/Slider)
 
 ## Multithumb support
