+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/base-fields/Slider/properties.mdx
