+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/base-fields/Upload.mdx
@@ -1,6 +1,7 @@
 ---
 title: 'Field.Upload'
 description: '`Field.Upload` is a wrapper for the Upload component to make it easier to use inside a form.'
+componentType: 'base-input'
 showTabs: true
 theme: ['sbanken', 'carnegie']
 tabs:
@@ -15,16 +16,14 @@ tabs:
 breadcrumb:
   - text: Forms
     href: /uilib/extensions/forms/
-  - text: Feature fields
-    href: /uilib/extensions/forms/feature-fields/
-  - text: More feature fields...
-    href: /uilib/extensions/forms/feature-fields/more-fields/
+  - text: Base fields
+    href: /uilib/extensions/forms/base-fields/
   - text: Upload
-    href: /uilib/extensions/forms/feature-fields/more-fields/Upload/
+    href: /uilib/extensions/forms/base-fields/Upload/
 ---
 
-import Info from 'Docs/uilib/extensions/forms/feature-fields/more-fields/Upload/info'
-import Demos from 'Docs/uilib/extensions/forms/feature-fields/more-fields/Upload/demos'
+import Info from 'Docs/uilib/extensions/forms/base-fields/Upload/info'
+import Demos from 'Docs/uilib/extensions/forms/base-fields/Upload/demos'
 
 <Info />
 <Demos />
