+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/base-fields/Upload/info.mdx
@@ -18,7 +18,7 @@ There is a corresponding [Value.Upload](/uilib/extensions/forms/Value/Upload) co
 ## Relevant links
 
 - [Source code](https://github.com/dnbexperience/eufemia/tree/main/packages/dnb-eufemia/src/extensions/forms/Field/Upload)
-- [Docs code](https://github.com/dnbexperience/eufemia/tree/main/packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/feature-fields/more-fields/Upload)
+- [Docs code](https://github.com/dnbexperience/eufemia/tree/main/packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/base-fields/Upload)
 
 ## The data and file format
 
@@ -82,7 +82,7 @@ The `onChange` event handler will return an array with [file item objects](/uili
 
 For error handling of invalid files, you can refer to the [Upload](/uilib/components/upload/) component for more details.
 
-Here is [an example](/uilib/extensions/forms/feature-fields/more-fields/Upload/#with-file-size-validation) of how to use the `fileHandler` property to validate file sizes.
+Here is [an example](/uilib/extensions/forms/base-fields/Upload/#with-file-size-validation) of how to use the `fileHandler` property to validate file sizes.
 
 ## About the `value` and `path` property
