+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/feature-fields/Password.mdx
@@ -17,14 +17,12 @@ breadcrumb:
     href: /uilib/extensions/forms/
   - text: Feature fields
     href: /uilib/extensions/forms/feature-fields/
-  - text: More feature fields...
-    href: /uilib/extensions/forms/feature-fields/more-fields/
   - text: Password
-    href: /uilib/extensions/forms/feature-fields/more-fields/Password/
+    href: /uilib/extensions/forms/feature-fields/Password/
 ---
 
-import Info from 'Docs/uilib/extensions/forms/feature-fields/more-fields/Password/info'
-import Demos from 'Docs/uilib/extensions/forms/feature-fields/more-fields/Password/demos'
+import Info from 'Docs/uilib/extensions/forms/feature-fields/Password/info'
+import Demos from 'Docs/uilib/extensions/forms/feature-fields/Password/demos'
 
 <Info />
 <Demos />
