+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/feature-fields/Password/Examples.tsx
@@ -1,4 +1,4 @@
-import ComponentBox from '../../../../../../../shared/tags/ComponentBox'
+import ComponentBox from '../../../../../../shared/tags/ComponentBox'
 import { Field } from '@dnb/eufemia/src/extensions/forms'
 
 export const Placeholder = () => {
