+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/feature-fields/Password/info.mdx
@@ -18,7 +18,7 @@ It supports the HTML [autocomplete](https://developer.mozilla.org/en-US/docs/Web
 ## Relevant links
 
 - [Source code](https://github.com/dnbexperience/eufemia/tree/main/packages/dnb-eufemia/src/extensions/forms/Field/Password)
-- [Docs code](https://github.com/dnbexperience/eufemia/tree/main/packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/feature-fields/more-fields/Password)
+- [Docs code](https://github.com/dnbexperience/eufemia/tree/main/packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/feature-fields/Password)
 
 ## Validation
