--- packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/feature-fields/more-fields.mdx
@@ -1,20 +1,0 @@
----
-title: 'more fields...'
-order: -1
-breadcrumb:
-  - text: Forms
-    href: /uilib/extensions/forms/
-  - text: Feature fields
-    href: /uilib/extensions/forms/feature-fields/
-  - text: More feature fields...
-    href: /uilib/extensions/forms/feature-fields/more-fields/
-accordion: true
----
-
-import ListMoreFieldComponents from './more-fields/ListMoreFieldComponents'
-
-# More feature fields
-
-Here is also a list of [all available fields](/uilib/extensions/forms/all-fields/).
-
-<ListMoreFieldComponents size="medium" />
