--- packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/feature-fields/more-fields/ListMoreFieldComponents.tsx
@@ -1,43 +1,0 @@
-import { useStaticQuery, graphql } from 'portal-query'
-import ListSummaryFromEdges from '../../../../../../shared/parts/ListSummaryFromEdges'
-
-export default function ListMoreFieldComponents(props) {
-  const {
-    allMdx: { edges },
-  } = useStaticQuery(graphql`
-    {
-      allMdx(
-        filter: {
-          frontmatter: {
-            showTabs: { ne: null }
-            title: { ne: null }
-            draft: { ne: true }
-          }
-          internal: {
-            contentFilePath: {
-              glob: "**/uilib/extensions/forms/feature-fields/more-fields/*"
-            }
-          }
-        }
-        sort: [
-          { frontmatter: { order: ASC } }
-          { frontmatter: { title: ASC } }
-        ]
-      ) {
-        edges {
-          node {
-            fields {
-              slug
-            }
-            frontmatter {
-              title
-              description
-            }
-          }
-        }
-      }
-    }
-  `)
-
-  return <ListSummaryFromEdges edges={edges} {...props} />
-}
